tv-program/requirements.txt
Pillow==9.1.0
tk==0.1.0
selenium==4.0.0

